Slightly refactored Targeted Skills handlers
