+++ src/GameLogic/PlayerActions/Skills/NovaSkillStartAction.cs
@@ -1,4 +1,4 @@
-﻿// <copyright file="NovaSkillAction.cs" company="MUnique">
+﻿// <copyright file="NovaSkillStartAction.cs" company="MUnique">
 // Licensed under the MIT License. See LICENSE file in the project root for full license information.
 // </copyright>
 
@@ -11,12 +11,10 @@ using MUnique.OpenMU.GameLogic.NPC;
 using MUnique.OpenMU.GameLogic.Views.World;
 
 /// <summary>
-/// The nova skill action.
+/// The start nova skill action.
 /// </summary>
-public class NovaSkillAction
+public class NovaSkillStartAction : TargetedSkillActionBase
 {
-    private const byte NovaSkillId = 40;
-    private const byte NovaStartId = 58;
     private static readonly TimeSpan NovaStepDelay = TimeSpan.FromMilliseconds(500);
 
     /// <summary>
@@ -24,49 +22,39 @@ public class NovaSkillAction
     /// </summary>
     private static readonly int[] NovaDamageTable = { 0, 20, 50, 99, 160, 225, 325, 425, 550, 700, 880, 1090, 1320 };
 
-    /// <summary>
-    /// Starts the nova skill.
-    /// </summary>
-    /// <param name="player">The player which performs the skill.</param>
-    public async ValueTask StartNovaSkillAsync(Player player)
+    /// <inheritdoc/>
+    public override ushort Key => 58;
+
+    /// <inheritdoc />
+    public override async ValueTask PerformSkillAsync(Player player, IAttackable target, ushort skillId)
     {
-        if (player.NovaCancellationTokenSource is not null)
+        if (player.TargetedSkillCancelTokenSource is not null)
         {
             // A nova is already ongoing.
             return;
         }
 
-        var skillEntry = player.SkillList?.GetSkill(NovaSkillId);
+        var skillEntry = player.SkillList?.GetSkill(skillId);
         if (skillEntry?.Skill is null)
         {
             return;
         }
 
         // Consume full ability points first...
-        var novaStart = player.GameContext.Configuration.Skills.First(s => s.Number == NovaStartId);
+        var novaStart = player.GameContext.Configuration.Skills.First(s => s.Number == skillId);
         if (!await player.TryConsumeForSkillAsync(novaStart).ConfigureAwait(false))
         {
             return;
         }
 
         await player.ForEachWorldObserverAsync<IShowSkillAnimationPlugIn>(p => p.ShowNovaStartAsync(player), true).ConfigureAwait(false);
-        var cancellationTokenSource = new NovaCancellationTokenSource();
-        player.NovaCancellationTokenSource = cancellationTokenSource;
+        var cancellationTokenSource = new TargetedSkillCancellationTokenSource();
+        player.TargetedSkillCancelTokenSource = cancellationTokenSource;
 
         _ = this.RunNovaAsync(player, skillEntry, cancellationTokenSource);
     }
 
-    /// <summary>
-    /// Stops the nova skill.
-    /// </summary>
-    /// <param name="player">The player which performed the skill.</param>
-    /// <param name="extraTargetId">The extra target identifier.</param>
-    public async ValueTask StopNovaSkillAsync(Player player, ushort extraTargetId)
-    {
-        player.NovaCancellationTokenSource?.CancelWithExtraTarget(extraTargetId);
-    }
-
-    private async ValueTask RunNovaAsync(Player player, SkillEntry skillEntry, NovaCancellationTokenSource cancellationTokenSource)
+    private async ValueTask RunNovaAsync(Player player, SkillEntry skillEntry, TargetedSkillCancellationTokenSource cancellationTokenSource)
     {
         var cancellationToken = cancellationTokenSource.Token;
         if (player.Attributes is not { } playerAttributes
@@ -95,7 +83,8 @@ public class NovaSkillAction
                     completedSteps++;
                     stepDamageElement.Value = NovaDamageTable[completedSteps];
                     var steps = completedSteps;
-                    await player.ForEachWorldObserverAsync<IShowSkillStageUpdatePlugIn>(p => p.UpdateSkillStageAsync(player, NovaSkillId, steps), true).ConfigureAwait(false);
+                    ushort skillId = this.Key;
+                    await player.ForEachWorldObserverAsync<IShowSkillStageUpdatePlugIn>(p => p.UpdateSkillStageAsync(player, skillId, steps), true).ConfigureAwait(false);
                     await Task.Delay(NovaStepDelay, cancellationToken).ConfigureAwait(false); // Hint: Player could cancel the nova 500 ms before end without damage loss - if he is good
                 }
             }
@@ -109,8 +98,8 @@ public class NovaSkillAction
         catch (Exception ex)
         {
             player.Logger.LogError(ex, "Unexpected error during performing nova skill");
-            player.NovaCancellationTokenSource?.Dispose();
-            player.NovaCancellationTokenSource = null;
+            player.TargetedSkillCancelTokenSource?.Dispose();
+            player.TargetedSkillCancelTokenSource = null;
         }
         finally
         {
@@ -118,7 +107,7 @@ public class NovaSkillAction
         }
     }
 
-    private async ValueTask AttackTargetsAsync(Player player, SkillEntry skillEntry, NovaCancellationTokenSource cancellationTokenSource)
+    private async ValueTask AttackTargetsAsync(Player player, SkillEntry skillEntry, TargetedSkillCancellationTokenSource cancellationTokenSource)
     {
         if (!player.IsAlive || player.IsAtSafezone() || skillEntry.Skill is not { } skill)
         {
@@ -130,8 +119,8 @@ public class NovaSkillAction
         var targets = this.DetermineTargets(player, skill, explicitTargetId);
 
         // Set cancellation token source to null, so that the next nova can be started.
-        player.NovaCancellationTokenSource?.Dispose();
-        player.NovaCancellationTokenSource = null;
+        player.TargetedSkillCancelTokenSource?.Dispose();
+        player.TargetedSkillCancelTokenSource = null;
 
         await Task.Delay(500, CancellationToken.None).ConfigureAwait(false);
