+++ src/GameLogic/PlayerActions/Skills/TargetedSkillCancellationTokenSource.cs
@@ -1,4 +1,4 @@
-﻿// <copyright file="NovaCancellationTokenSource.cs" company="MUnique">
+﻿// <copyright file="TargetedSkillCancelTokenSource.cs" company="MUnique">
 // Licensed under the MIT License. See LICENSE file in the project root for full license information.
 // </copyright>
 
@@ -10,7 +10,7 @@ using System.Threading;
 /// A <see cref="CancellationTokenSource"/> which allows to specify an explicit target when cancelling the nova skill.
 /// </summary>
 /// <seealso cref="System.Threading.CancellationTokenSource" />
-public class NovaCancellationTokenSource : CancellationTokenSource
+public class TargetedSkillCancellationTokenSource : CancellationTokenSource
 {
     /// <summary>
     /// Gets the explicit target identifier.
